requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
pyasn1==0.4.8
pyjsbn-rsa==0.4.2
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2
urllib3==1.26.3
